+++ spec/syntax/versions/aliases_spec.rb
@@ -25,7 +25,9 @@ RSpec.describe(Regexp::Syntax) do
   include_examples 'syntax alias', 'ruby/2.5.0',  Regexp::Syntax::V2_4_1
   include_examples 'syntax alias', 'ruby/2.5',    Regexp::Syntax::V2_5_0
   include_examples 'syntax alias', 'ruby/2.6.0',  Regexp::Syntax::V2_5_0
-  include_examples 'syntax alias', 'ruby/2.6',    Regexp::Syntax::V2_5_0
+  include_examples 'syntax alias', 'ruby/2.6.2',  Regexp::Syntax::V2_6_2
+  include_examples 'syntax alias', 'ruby/2.6.3',  Regexp::Syntax::V2_6_3
+  include_examples 'syntax alias', 'ruby/2.6',    Regexp::Syntax::V2_6_3
 
   specify('future alias warning') do
     expect { Regexp::Syntax.new('ruby/5.0') }